Move server code to src/appengine directory
