+++ .gitignore
@@ -100,5 +100,5 @@ public/
 
 # End of https://www.gitignore.io/api/node,appengine
 
-*.js
+src/**/*.js
 .env
