+++ src/appengine/utils/stripTrailingSlash.ts
@@ -1,0 +1,5 @@
+/**
+ * Removes the trailing slash char.
+ */
+export default (str: string): string =>
+  str.endsWith("/") ? str.slice(0, -1) : str
